Camera component (#3)
* Camera added

tweaked some artist values
updated the texture collection to not lazy load, because I don't see why I would want to use lazy loading for it, honestly.
Should be able to move the camera at this point and everything look good. (Except the cursor)

* janky progress

some scrolling is working, but not quite all there yet.

* some conversions

shift the camera to work on doubled coordinates. Need to shift the artist as well, I think.

* fixeded

doing most of the math in doubled coordinates. Improving the drawing by a ton. The camera scrolls properly and the cursor is moving to the proper position!

* cleanup