Centering the camera (#4)
* starting work

centering is mostly there, but borked, it looks like either the way the offset is calculated or the fact that we end up with horizontal offset values at (literally) odd values and it was not designed with that in mind.

* constant, forever refactoring

* Oh hey, I figured it out

also commented out the text that pops up on screen. I don't need that right now.

* adding test stuff

* camera & mapArtist update

Well, we've completed map centering. Should probably do more refactoring and then start testing. Or I could close this branch and start a testing branch to do the refactoring in. Probably a good call.