Breaking stuff probably (#6)
* added a game menu state

* That's not right

Putting the unit list into the player class really isn't that great of an idea, is it?

* shifting artist to patron

* unit changed from player-based to tile-based

* Update MapPatron.cs

So you can actually see a unit. You know, silly things like that.

* started work on movement range

still not entirely sure about how to go about setting up the results

* Dijkstra mostly set up.

The basics are done. Need to set up unit movement limits and movement based on terrain.