Cleanup the scenario class (#7)
* more updates

* Update CursorPlayState.cs

* clearing out more stuff

just trying to use the scenario class instead of the map/camera/cursor/map objects.

* I don't think I broke anything

* Update UnitSelectedState.cs

Yeah, it was broken because the first value was not set. So if you selected a unit, you'd get a null reference exception. Not good. But it's fixed now, so it's all good now.

* messing with cursor

* replacing Monogame Point

Built new object for keeping track of position of units and tiles. Can easily shift between Doubled Offset and Cubed/Axial.